requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
exceptiongroup==1.1.2
fastapi==0.100.0
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
multidict==6.0.4
pydantic==2.1.1
pydantic_core==2.4.0
PyMySQL==1.1.0
PyYAML==6.0.1
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.1
yarl==1.9.2
